"hyosangs_sonar" defect and review fixes reflect work.
This block of commented-out lines of code should be removed.
#200 #201

The type of the object should be an interface such as "List" rather than the implementation "LinkedList" or "ArrayList"
#345 #343 #342 #341 #335 #328

Replace the type specification in this constructor call with the diamond operator ("<>"). (sonar.java.source not set. Assuming 7 or greater.)
#346 #344 #340 #333 #331 #329 #307 #303 #302 #301 #295 #284 #251

Convert this Map to an EnumMap
#332 #330

Override "equals(Object obj)" to comply with the contract of the "compareTo(T o)" method.
#337 #308

Remove this unnecessary cast to "Inspectable".
#223

Replace the synchronized class "Vector" by an unsynchronized one such as "ArrayList" or "LinkedList"
#198 #197 #196 #195 #194 #193 #192 #191 #190 #184 #173 #168 #166 #165 #164 #163 #146 #131 #65 #63 #62

code review f/up
isEmpty => Use isEmpty() to check whether the collection is empty or not
(#281 #280 #279 #272)

CompareTo => Override "equals(Object obj)" to comply with the contract of the "compareTo(T o)" method. (#337 #308)

Rename "goals" which hides the field declared at line 56.
src/edu/cmu/tartan/Game.java L596
#130